% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{escape_space_glue}
\alias{escape_space_glue}
\title{escape_space_glue}
\usage{
escape_space_glue(string, participants_attributes)
}
\arguments{
\item{string}{\dots}

\item{participants_attributes}{\dots}
}
\description{
escape_space_glue
}
\keyword{internal}
