% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mail_delivery_failure}
\alias{mail_delivery_failure}
\title{mail_delivery_failure}
\usage{
mail_delivery_failure(
  sqlite_base,
  imap_server,
  username,
  password,
  mailbox,
  delete = FALSE
)
}
\arguments{
\item{sqlite_base}{\dots}

\item{imap_server}{\dots}

\item{username}{\dots}

\item{password}{\dots}

\item{mailbox}{\dots}

\item{delete}{\dots}
}
\description{
mail_delivery_failure
}
