% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mailing}
\alias{mailing}
\title{mailing}
\usage{
mailing(
  rv,
  participants,
  from,
  subject,
  body,
  sleep,
  delete_survey = TRUE,
  general = FALSE
)
}
\arguments{
\item{rv}{\dots}

\item{participants}{\dots}

\item{from}{\dots}

\item{subject}{\dots}

\item{body}{\dots}

\item{sleep}{\dots}

\item{delete_survey}{\dots}

\item{general}{\dots}
}
\description{
mailing
}
\keyword{internal}
