% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_config_api.R
\name{mod_config_api_ui}
\alias{mod_config_api_ui}
\alias{mod_config_api_server}
\title{mod_config_api_ui and mod_config_api_server}
\usage{
mod_config_api_ui(id)

mod_config_api_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
