% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_crowdsourcing.R
\name{mod_crowdsourcing_ui}
\alias{mod_crowdsourcing_ui}
\alias{mod_crowdsourcing_server}
\title{mod_crowdsourcing_ui and mod_crowdsourcing_server}
\usage{
mod_crowdsourcing_ui(id)

mod_crowdsourcing_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
