% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_email_validation.R
\name{mod_email_validation_ui}
\alias{mod_email_validation_ui}
\alias{mod_email_validation_server}
\title{mod_email_validation_ui and mod_email_validation_server}
\usage{
mod_email_validation_ui(id)

mod_email_validation_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
