% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import_contacts.R
\name{mod_import_contacts_ui}
\alias{mod_import_contacts_ui}
\alias{mod_import_contacts_server}
\title{mod_import_contacts_ui and mod_import_contacts_server}
\usage{
mod_import_contacts_ui(id)

mod_import_contacts_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
