% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_incomplete_responses.R
\name{mod_incomplete_responses_ui}
\alias{mod_incomplete_responses_ui}
\alias{mod_incomplete_responses_server}
\title{mod_incomplete_responses_ui and mod_incomplete_responses_server}
\usage{
mod_incomplete_responses_ui(id)

mod_incomplete_responses_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
