% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_linkedin.R
\name{mod_linkedin_ui}
\alias{mod_linkedin_ui}
\alias{mod_linkedin_server}
\title{mod_linkedin_ui and mod_linkedin_server}
\usage{
mod_linkedin_ui(id)

mod_linkedin_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
