% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mailing.R
\name{mod_mailing_ui}
\alias{mod_mailing_ui}
\alias{mod_mailing_server}
\title{mod_mailing_ui and mod_mailing_server}
\usage{
mod_mailing_ui(id)

mod_mailing_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
