% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_participants.R
\name{mod_participants_ui}
\alias{mod_participants_ui}
\alias{mod_participants_server}
\title{mod_participants_ui and mod_participants_server}
\usage{
mod_participants_ui(id)

mod_participants_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{rv}{internal}

\item{get_filter_surveys}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
