% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_phoning.R
\name{mod_phoning_ui}
\alias{mod_phoning_ui}
\alias{mod_phoning_server}
\title{mod_phoning_ui and mod_phoning_server}
\usage{
mod_phoning_ui(id)

mod_phoning_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
