% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sms_sending.R
\name{mod_sms_sending_ui}
\alias{mod_sms_sending_ui}
\alias{mod_sms_sending_server}
\title{mod_sms_sending_ui and mod_sms_sending_server}
\usage{
mod_sms_sending_ui(id)

mod_sms_sending_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
