% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_surveys.R
\name{mod_surveys_ui}
\alias{mod_surveys_ui}
\alias{mod_surveys_server}
\title{mod_import_surveys_ui and mod_import_surveys_server}
\usage{
mod_surveys_ui(id)

mod_surveys_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
