% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pours_etud_perte_reprise}
\alias{pours_etud_perte_reprise}
\title{pours_etud_perte_reprise}
\usage{
pours_etud_perte_reprise(sqlite_base)
}
\arguments{
\item{sqlite_base}{\dots}
}
\description{
pours_etud_perte_reprise
}
