% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_finished_almost_complete}
\alias{set_finished_almost_complete}
\title{set_finished_almost_complete}
\usage{
set_finished_almost_complete(
  sqlite_base,
  cron_responses,
  almost_complete_group = c(`123459` = 16, `123458` = 16, `123456` = 16),
  token = NULL,
  force_today = FALSE
)
}
\arguments{
\item{sqlite_base}{\dots}

\item{cron_responses}{\dots}

\item{almost_complete_group}{\dots}

\item{token}{\dots}
}
\description{
set_finished_almost_complete
}
